Allow cf-deployment version to float
This is a very hacky solution. In order to reuse the same pipeline file
for both floating and locked cf-deployment pipelines, we take advantage
of the fact that resource source configuration ignores unknown keys.

We introduce the "set_to_tag_filter_to_lock_cf_deployment" variable in
place of the "tag_filter" key. When that variable is set to
"tag_filter", then the resource behaves as before and locks to version
"0.0.0". When the variable is set to anything else, the key is ignored,
and no tag_filtering occurs.

[#136066129]